requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
bcrypt>=3.2.0
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
python-multipart>=0.0.5
apscheduler>=3.8.0
aiohttp>=3.8.0
cryptography>=36.0.0
pillow>=9.0.0
pydantic>=1.9.0